refactor some code & fix shade spawn size for same room
